move KeyedOperator to operator package
