+++ generate/src/main/kotlin/uk/dioxic/helios/generate/operators/KeyedOperator.kt
@@ -1,4 +1,6 @@
-package uk.dioxic.helios.generate
+package uk.dioxic.helios.generate.operators
+
+import uk.dioxic.helios.generate.Operator
 
 abstract class KeyedOperator<T> : Operator<T> {
     abstract val key: String
